Add visualization

+++ lab1/filesystem.hpp
@@ -7,15 +7,15 @@
 #include <vector>
 
 namespace utils {
-std::stringstream read(std::ifstream &in) {
+std::stringstream read(std::ifstream& in) {
     std::stringstream sstr;
     sstr << in.rdbuf();
     return sstr;
 }
 
 void read_from_file(
-    std::string m1_filename, std::string m2_filename, std::stringstream &m1,
-    std::stringstream &m2) {
+    std::string m1_filename, std::string m2_filename, std::stringstream& m1,
+    std::stringstream& m2) {
     std::ifstream file(m1_filename);
     m1 = read(file);
     file.close();
@@ -24,9 +24,9 @@ void read_from_file(
     file.close();
 }
 
-void fill_matrix(std::stringstream &buf, std::vector<uint32_t> &vec) {
+    void fill_matrix(std::stringstream& buf, std::vector<uint32_t>& vec) {
     std::string token;
-    uint64_t size, n, m;
+    size_t size, n, m;
     size = m = n = 0;
     while(buf >> token) {
         switch(size) {
@@ -40,11 +40,11 @@ void fill_matrix(std::stringstream &buf, std::vector<uint32_t> &vec) {
             break;
         case 2:
             vec.reserve(n * m + 2);
-            vec.push_back(stod(token));
+            vec.push_back(stoul(token));
             ++size;
             break;
         default:
-            vec.push_back(stod(token));
+            vec.push_back(stoul(token));
             break;
         }
     }
